fix(plugins): satisfy uninstall cleanup checks
